+++ crates/nostr/src/nips/nip35.rs
@@ -77,6 +77,8 @@ impl Torrent {
             tags.push(Tag::hashtag(tag));
         }
 
-        EventBuilder::new(Kind::Torrent, self.description).tags(tags)
+        EventBuilder::new(Kind::Torrent)
+            .content(self.description)
+            .tags(tags)
     }
 }