+++ crates/nostr/src/nips/nipb0.rs
@@ -89,6 +89,8 @@ impl WebBookmark {
             tags.push(TagStandard::Hashtag(hashtag).into());
         }
 
-        EventBuilder::new(Kind::WebBookmark, self.description).tags(tags)
+        EventBuilder::new(Kind::WebBookmark)
+            .content(self.description)
+            .tags(tags)
     }
 }