+++ signer/nostr-browser-signer-proxy/examples/browser-signer-proxy.rs
@@ -38,7 +38,9 @@ async fn main() -> Result<()> {
     // Build a gift wrap
     let receiver =
         PublicKey::parse("npub1drvpzev3syqt0kjrls50050uzf25gehpz9vgdw08hvex7e0vgfeq0eseet")?;
-    let rumor = EventBuilder::new(Kind::Custom(123), "test").build(public_key);
+    let rumor = EventBuilder::new(Kind::Custom(123))
+        .content("test")
+        .build(public_key);
     let gift_wrap = EventBuilder::gift_wrap(&proxy, &receiver, rumor, []).await?;
     println!("Gift wrap: {}", gift_wrap.as_json());